+++ NEWS.rst
@@ -6,7 +6,7 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Minor: Added support for "empty" metrics object with a count of 0.
 
 4.1.0
 -----